requirements.txt
langchain>=0.2.10
langchain-community>=0.2.9
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0

sentence-transformers>=3.0.1
torch>=2.2.0

chromadb>=1.1.0
faiss-cpu>=1.8.0
FlagEmbedding>=1.3.0

beautifulsoup4>=4.12.3
trafilatura>=2.0.0
markdownify>=0.12.1

pyyaml>=6.0.1
requests>=2.32.2
python-slugify>=8.0.4
pydantic>=2.7.1

fastapi>=0.115.0
uvicorn>=0.30.0